Reorganized project structure into separate folders for each grouping of routes
