+++ app/main/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+# Creates a blueprint for all our code under main
+main = Blueprint('main', __name__)
+
+from app.main import views
