+++ app/main/views.py
@@ -1,0 +1,7 @@
+from . import main
+
+
+@main.route('/')
+def hello_world():
+    main.logger.debug("TESTING DEBUG LOG FUNCTIONALITY")
+    return "Hello, world!"
